# Optimize `MatchNotRegexp` and `MatchNotEqual`